Rename sample data file
